Migrate RobolectricAttributeSetTest to use new Robolectric.buildAttributeSet() API, remove any tests that didn't make sense, e.g: creating in different name spaces doesn't make sense since all R values are reconciled to the same ID space.
Move all its methods to new AttributeSetBuilderTest.

mark RoboAttributeSet as @Deprecated, will be removed in next release.
